+++ src/coverConfig.ts
